feat(tasks): update error handling for db query

+++ app/api/tasks/[id]/route.ts
@@ -21,31 +21,42 @@ export async function GET(
     );
   }
 
-  const { id } = await ctx.params;
-  const task = await prisma.task.findFirst({
-    where: {
-      id: Number(id),
-      userId: userId,
-    },
-  });
+  try {
+    const { id } = await ctx.params;
+    const task = await prisma.task.findFirst({
+      where: {
+        id: Number(id),
+        userId: userId,
+      },
+    });
 
-  // not found error response
-  if (!task) {
+    // not found error response
+    if (!task) {
+      return NextResponse.json(
+        { status: "error", message: "Task not found." },
+        { status: 404 },
+      );
+    }
+
+    // success response
     return NextResponse.json(
-      { status: "error", message: "Task not found." },
-      { status: 404 },
+      {
+        status: "success",
+        message: `Berhasil mengambil tugas ${task?.id}`,
+        data: task,
+      },
+      { status: 200 },
+    );
+  } catch (error) {
+    return NextResponse.json(
+      {
+        status: "error",
+        message: "Something went wrong.",
+        data: null,
+      },
+      { status: 500 },
     );
   }
-
-  // success response
-  return NextResponse.json(
-    {
-      status: "success",
-      message: `Berhasil mengambil tugas ${task?.id}`,
-      data: task,
-    },
-    { status: 200 },
-  );
 }
 
 export async function PATCH(
@@ -95,26 +106,37 @@ export async function PATCH(
     );
   }
 
-  const task = await prisma.task.update({
-    where: {
-      id: Number(id),
-    },
-    data: {
-      title: parsed.data.title ?? existingTask.title,
-      description: parsed.data.description ?? existingTask.description,
-      status: parsed.data.status ?? existingTask.status,
-    },
-  });
+  try {
+    const task = await prisma.task.update({
+      where: {
+        id: Number(id),
+      },
+      data: {
+        title: parsed.data.title ?? existingTask.title,
+        description: parsed.data.description ?? existingTask.description,
+        status: parsed.data.status ?? existingTask.status,
+      },
+    });
 
-  // success response
-  return NextResponse.json(
-    {
-      status: "success",
-      message: `Berhasil mengubah tugas ${task?.id}`,
-      data: task,
-    },
-    { status: 200 },
-  );
+    // success response
+    return NextResponse.json(
+      {
+        status: "success",
+        message: `Berhasil mengubah tugas ${task?.id}`,
+        data: task,
+      },
+      { status: 200 },
+    );
+  } catch (error) {
+    return NextResponse.json(
+      {
+        status: "error",
+        message: "Something went wrong.",
+        data: null,
+      },
+      { status: 500 },
+    );
+  }
 }
 
 export async function DELETE(
@@ -135,21 +157,45 @@ export async function DELETE(
   }
 
   const { id } = await ctx.params;
-
-  const deletedTask = await prisma.task.delete({
+  const existingTask = await prisma.task.findUnique({
     where: {
       id: Number(id),
-      userId: userId,
     },
   });
 
-  // success response
-  return NextResponse.json(
-    {
-      status: "success",
-      message: `Berhasil menghapus tugas ${deletedTask.id}`,
-      data: deletedTask,
-    },
-    { status: 200 },
-  );
+  // not found error response
+  if (!existingTask) {
+    return NextResponse.json(
+      { status: "error", message: "Task not found" },
+      { status: 404 },
+    );
+  }
+
+  try {
+    const deletedTask = await prisma.task.delete({
+      where: {
+        id: Number(id),
+        userId: userId,
+      },
+    });
+
+    // success response
+    return NextResponse.json(
+      {
+        status: "success",
+        message: `Berhasil menghapus tugas ${deletedTask.id}`,
+        data: deletedTask,
+      },
+      { status: 200 },
+    );
+  } catch (error) {
+    return NextResponse.json(
+      {
+        status: "error",
+        message: "Something went wrong.",
+        data: null,
+      },
+      { status: 500 },
+    );
+  }
 }
